backend/requirements.txt
APScheduler==3.11.0
Flask==3.1.0
flask_cors==5.0.1
flask_sqlalchemy==3.1.1
python-dotenv==1.1.0
schedule==1.2.2
selenium==4.31.0
webdriver_manager==4.0.2
